Split renderables into individual files (#29)
